# cmd/api/openapi/openapi.yaml
openapi: 3.1.0

info:
  title: RuangWali API
  version: 1.0.0
  description: |
    API backend RuangWali untuk autentikasi, session management,
    access control, dan layanan akademik.

  contact:
    name: RuangWali Development Team

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
    description: Application health checks

  - name: Authentication
    description: Authentication and token lifecycle

  - name: Password
    description: Password management and password recovery

  - name: Sessions
    description: User refresh-session management

paths:
  /health/live:
    get:
      tags:
        - Health
      summary: Liveness check
      operationId: getLiveness
      responses:
        "200":
          description: Application process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /health/ready:
    get:
      tags:
        - Health
      summary: Readiness check
      operationId: getReadiness
      responses:
        "200":
          description: Application is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"

        "503":
          description: Application is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"

  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logout successful

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/change-password:
    post:
      tags:
        - Password
      summary: Change authenticated user password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed successfully

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/forgot-password:
    post:
      tags:
        - Password
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Password reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/reset-password:
    post:
      tags:
        - Password
      summary: Reset password using reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset successful

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/sessions:
    get:
      tags:
        - Sessions
      summary: List authenticated user sessions
      operationId: listSessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Sessions
      summary: Revoke all user sessions
      operationId: revokeAllSessions
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeAllSessionsRequest"
      responses:
        "204":
          description: Sessions revoked successfully

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/sessions/{sessionID}:
    delete:
      tags:
        - Sessions
      summary: Revoke specific user session
      operationId: revokeSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          description: Refresh session UUID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session revoked successfully

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@ruangwali.local

        password:
          type: string
          format: password
          minLength: 1
          example: RuangWali@2026

    LogoutRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          minLength: 1

    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          minLength: 1

    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 1

        newPassword:
          type: string
          format: password
          minLength: 8
          example: NewRuangWali@2026

    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: admin@ruangwali.local

    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          minLength: 1

        newPassword:
          type: string
          format: password
          minLength: 8
          example: NewRuangWali@2026

    RevokeAllSessionsRequest:
      type: object
      additionalProperties: false
      properties:
        keepCurrent:
          type: boolean
          default: false

    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string

        refreshToken:
          type: string

        tokenType:
          type: string
          example: Bearer

        expiresIn:
          type: integer
          format: int64
          minimum: 0
          description: Access token lifetime in seconds

        user:
          $ref: "#/components/schemas/UserResponse"

    CurrentUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"

    UserResponse:
      type: object
      required:
        - id
        - email
        - status
      properties:
        id:
          type: string
          format: uuid

        email:
          type: string
          format: email

        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - SUSPENDED

        lastLoginAt:
          type:
            - string
            - "null"
          format: date-time

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time

    SessionResponse:
      type: object
      required:
        - id
        - userId
        - familyId
        - expiresAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid

        userId:
          type: string
          format: uuid

        familyId:
          type: string
          format: uuid

        userAgent:
          type: string

        ipAddress:
          type: string

        expiresAt:
          type: string
          format: date-time

        lastUsedAt:
          type:
            - string
            - "null"
          format: date-time

        revokedAt:
          type:
            - string
            - "null"
          format: date-time

        revokeReason:
          type:
            - string
            - "null"

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time

    SessionListResponse:
      type: object
      required:
        - sessions
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionResponse"

    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: invalid_request

        message:
          type: string
          example: request tidak valid

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    ReadinessResponse:
      type: object
      required:
        - status
        - database
        - build
      properties:
        status:
          type: string
          example: ready

        database:
          type: string
          example: available

        build:
          $ref: "#/components/schemas/BuildInfo"

    BuildInfo:
      type: object
      additionalProperties: true